requirements.txt
@"
Django>=5.2,<5.3
python-dotenv>=1.0,<2.0
Pillow>=10.0
"@ | Set-Content .\requirements.txt